move introspect to it's own file
